Fixes eternal space traveling for items
Now anything that is not client-carrying mob is deleted when reaching edge of zlevel.
Space is big, and it stops random items travelling forever and ever, potentially creating more zlevels.
As a bonus, if you were pulling something when reaching edge, it carries over with you, so you can carry stuff to sectors via lucky EVA.